pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tramtrix"
version = "0.1.0"
description = "A simple tool to display Zurich tram departures on a Ulanzi clock running Awtrix 3."
readme = "README.md"
requires-python = ">=3.8"
authors = [
  {name = "Tom", email = "contact@tomm.org"},
]
dependencies = [
  "requests",
  "python-dotenv",
]

[project.scripts]
tramtrix = "tramtrix.main:main"
